Refactoring: InternalListChangeLister is private internal class
